Redesign prompt generation methods to support multi-phase approaches
The previous design for prompt generation generated a single prompt
based on the task  and combined this with a single prompt based on
the prompt mode. However, for any given task different prompts are
needed for different prompt modes, therefore we split the single
task prompt method into 4 prompt methods, replacing the
functionality of both single prompt methods (task_prompt and
i3_prompt) which were removed.