Refactor 'execute_api_calls' from single shot exec to interactive loop
Previously LLM API calls were implemented with training and testing
executed in a single shot but the ability to execute training and
testing separately was desirable, especially for the I3 prompting
approach. This is to allow the classification prompt generated in
the training phase to be modified by the user before testing and so
running a test on the same classification prompt could be executed
multiple times in a row. This fix changes 'execute_api_calls' to be
an interactive loop for the 'basic' and 'basic_with_context' prompt
design and adds execute_api_calls for the 'i3' prompt design. A
variety of small code changes were also made to accommodate this
design change.